Set the 'expformat' parameter of plotms if not allready provided by the user.
Recipe changes the 'plotfile' parameter of plotms to a random filename.
Plotms normally will determine the data format from the filename suffix,
but the randomly generated filename doesn't contain a suffix to determine
its type. In this case it is needed to set the 'expformat' parameter, and
manually specify the data format or plotms will fail (without an error
message).