Explicitly set default value of "interp".
A number of tasks take the "interp" parameter, unfortunately, the
the default value for "interp" changes between runs. So running the
same task twice with the same arguments will not always produce the
same hash! The default value for 'interp' will be either [""] or
["linear"], which are actually equivalent. As a work around we just
set interp to linear when it is not set.